added doDoubleTapSelection method, remove inline updateSelection code in uikit sourceset
